+++ src/main/java/com/mrh0/horth/output/instructions/high/stackops/binop/HSub.java
@@ -7,7 +7,7 @@ import com.mrh0.horth.typechecker.IContract;
 import com.mrh0.horth.typechecker.types.AllTypes;
 
 public class HSub extends HighInst {
-    private static Contract contract = new Contract.Builder().pop(AllTypes.INT64, AllTypes.INT64).push(AllTypes.INT64).build();
+    private static Contract contract = new Contract.Builder().pop(AllTypes.INT, AllTypes.INT).push(AllTypes.INT).build();
 
     public HSub(ITok token) {
         super(token);
